Made DQMFileSaver::fillJson a stand alone function
Since fillJson is used by DQMFileSaver and JsonWritingTimeoutPoolOutputModule
it is better to make it a stand alone function.
