+++ DQMServices/Components/plugins/DQMFileSaver.h
@@ -18,8 +18,6 @@ public:
   DQMFileSaver(const edm::ParameterSet &ps);
 
 
-  // used by the JsonWritingTimedPoolOutputModule,
-  // fms will be nullptr in such case
   static boost::property_tree::ptree fillJson(
       int run, int lumi, const std::string &dataFilePathName, const std::string& transferDestinationStr,
       const std::string& mergeTypeStr, evf::FastMonitoringService *fms);
